requirements.txt
pdfplumber>=0.9.0
flask>=2.3.0
google-generativeai>=0.3.0
qdrant-client>=1.17.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
